requirements.txt
streamlit
pymupdf
pillow
openai
instructor